requirements.txt


python-dotenv
livekit-agents
openai
elevenlabs
silero
